% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sensor_data}
\alias{sensor_data}
\title{sensor_data dataset}
\format{
A data frame with 672500 rows and 4 variables
}
\source{
Purpleair Air Quality Dataset Kaggle
}
\usage{
sensor_data
}
\description{
Tidy dataset for Shiny app analysis.
}
\details{
Variables:
* `coordinates_by _time` : latitude longitude
* `time` : time point
* `value` : air quality index
}
\keyword{datasets}
